refactor(collection-points): rename useCoordinatesFromUrl to useMapUrlParams and relocate implementation
